apps/api/requirements.txt
fastapi[standard]
chromadb
requests
bs4
python-dotenv
ast
openai
pandas